leehyunsoo/django_framework/requirements.txt
pip
setuptools
django
ipython
jupyter
djangorestframework
pygments
httpie
pytest
pytest-django
django-test-plus
pytest-cov
flake8
pyflakes
codecov
django-debug-toolbar
pytest-profiling
pytest-xdist
snakeviz
